Move some classes from pdfarrager to core
To make the pdfarranger module less bloated
